Niyaz

+++ backend/src/main/java/com/wedspot/backend/Model/VendorServiceDTO.java
@@ -2,9 +2,9 @@ package com.wedspot.backend.Model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.wedspot.backend.Model.Entity.Review;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -26,13 +26,13 @@ public class VendorServiceDTO {
 
     private int ratingCount;
 
-    private double price;
+    private BigDecimal price;
 
     private String location;
 
     private String category;
 
-    private double quantity;
+    private Integer quantity;
 
     private UserDTO vendor;
